bool and double fixes
